refactor(entityRoutes): Add entityEditorMarkup
This function encapsulates the markup requred for creation/editing
of all the entities.
